+++ packages/utils-dev-dependency/plaid/testData/index.ts
@@ -2,3 +2,4 @@ export * from "./items";
 export * from "./eventsWebhooks";
 export * from "./accounts";
 export * from "./identity";
+export * from "./transactions";